Refactor code and remove unused test file
